тест для модели товара

+++ products/models.py
@@ -20,7 +20,11 @@ class Product(models.Model):
         verbose_name="Дата выхода продукта на рынок",
         help_text="Введите дату выхода продукта на рынок",
     )
-    price = models.FloatField(default=0, verbose_name="Стоимость товара")
+    price = models.DecimalField(
+        max_digits=15,
+        decimal_places=2,
+        default=0.00,
+        verbose_name="Стоимость товара")
     owner = models.ForeignKey(User, on_delete=models.SET_NULL, **NULLABLE)
 
     def __str__(self):
